Added new allocator for QuadreeTree

+++ source/Creature.cpp
@@ -338,7 +338,7 @@ void Creature::doTurn()
     }
 
     // Check to see if we have earned enough experience to level up.
-    while (exp >= 5 * (getLevel() + powl(getLevel() / 3.0, 2)) && getLevel() < 100)
+    while (exp >= 5 * (getLevel() + pow(getLevel() / 3.0, 2)) && getLevel() < 100)
     {
         doLevelUp();
     }
@@ -429,7 +429,7 @@ void Creature::doTurn()
         }
 
         // If we have found a home tile to sleep on, see if we are tired enough to want to go to sleep.
-        if (homeTile != 0 && 100.0 * powl(Random::Double(0.0, 0.8), 2) > awakeness && peekAction().getType() != CreatureAction::sleep)
+        if (homeTile != 0 && 100.0 * pow(Random::Double(0.0, 0.8), 2) > awakeness && peekAction().getType() != CreatureAction::sleep)
         {
             tempAction.setType(CreatureAction::sleep);
             pushAction(tempAction);
@@ -1356,7 +1356,7 @@ void Creature::doTurn()
                     }
 
                     // Randomly decide to stop training, we are more likely to stop when we are tired.
-                    if (100.0 * powl(Random::Double(0.0, 1.0), 2) > awakeness)
+                    if (100.0 * pow(Random::Double(0.0, 1.0), 2) > awakeness)
                     {
                         popAction();
                         trainWait = 0;
@@ -1503,7 +1503,7 @@ void Creature::doTurn()
                         double damageDone = getHitroll(getGameMap()->crowDistance(
                                 myTile, tempTile));
                         damageDone *= Random::Double(0.0, 1.0);
-                        damageDone -= powl(Random::Double(0.0, 0.4), 2.0)
+                        damageDone -= pow(Random::Double(0.0, 0.4), 2.0)
                                 * tempAttackableObject->getDefense();
 
                         // Make sure the damage is positive.
@@ -1513,7 +1513,7 @@ void Creature::doTurn()
                         // Do the damage and award experience points to both creatures.
                         tempAttackableObject->takeDamage(damageDone, tempTile);
                         double expGained;
-                        expGained = 1.0 + 0.2 * powl(damageDone, 1.3);
+                        expGained = 1.0 + 0.2 * pow(damageDone, 1.3);
                         awakeness -= 0.5;
 
                         // Give a small amount of experince to the creature we hit.
@@ -1765,7 +1765,7 @@ void Creature::doLevelUp()
  */
 void Creature::updateVisibleTiles()
 {
-    //double effectiveRadius = min(5.0, sightRadius) + sightRadius*powl(randomDouble(0.0, 1.0), 3.0);
+    //double effectiveRadius = min(5.0, sightRadius) + sightRadius*pow(randomDouble(0.0, 1.0), 3.0);
     //double effectiveRadius = sightRadius;
     //visibleTiles = getGameMap()->visibleTiles(positionTile(), effectiveRadius);
     visibleTiles = getGameMap()->visibleTiles(positionTile(), definition->getSightRadius());
@@ -1856,7 +1856,7 @@ std::vector<GameEntity*> Creature::getEnemyObjectsInRange(
         Tile *tempTile = enemyObjectsToCheck[i]->getCoveredTiles()[0];
         if (tempTile != NULL)
         {
-            double rSquared = powl(myTile->x - tempTile->x, 2.0) + powl(
+            double rSquared = pow(myTile->x - tempTile->x, 2.0) + pow(
                     myTile->y - tempTile->y, 2.0);
 
             if (rSquared < weaponRangeSquared)
@@ -2198,7 +2198,7 @@ void Creature::computeBattlefield()
             Tile *tempTile2 = tempObject->getCoveredTiles()[0];
 
             // Compensate for how close the creature is to me
-            //rSquared = powl(myTile->x - tempTile2->x, 2.0) + powl(myTile->y - tempTile2->y, 2.0);
+            //rSquared = pow(myTile->x - tempTile2->x, 2.0) + pow(myTile->y - tempTile2->y, 2.0);
             //double factor = 1.0 / (sqrt(rSquared) + 1.0);
 
             // Subtract for the distance from the enemy creature to r
@@ -2215,7 +2215,7 @@ void Creature::computeBattlefield()
             Tile *tempTile2 = visibleAlliedObjects[j]->getCoveredTiles()[0];
 
             // Compensate for how close the creature is to me
-            //rSquared = powl(myTile->x - tempTile2->x, 2.0) + powl(myTile->y - tempTile2->y, 2.0);
+            //rSquared = pow(myTile->x - tempTile2->x, 2.0) + pow(myTile->y - tempTile2->y, 2.0);
             //double factor = 1.0 / (sqrt(rSquared) + 1.0);
 
             xDist = tempTile->x - tempTile2->x;
